half-assed API tests
using some horrible object stubbing. better tests are needed, but those really need to be done in a different way as integration tests, so we can do things like test responses when Redis is down, etc etc. That's going to require some thinking about architecture.  But this can be a basic starting point before then.